// src/lib/pricing-types.ts
export interface StorePayload {
  name: string;
  website_url?: string;
  logo_url?: string;
  is_active?: boolean;
}